requirements.txt
django
djangorestframework
coreapi
django-filter
drf_openapi
django-rest-swagger
pyfcm
xbee
python-telegram-bot
requests
